# docs/10-critical-remediation/api-reference-openapi.yaml
openapi: 3.1.0
info:
  title: Ubertruck MVP API Reference
  description: Complete REST API specification for Ubertruck Logistics Platform
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@ubertruck.in
  license:
    name: Proprietary
    url: https://ubertruck.in/terms
servers:
  - url: https://api.ubertruck.in/v1
    description: Production Server
  - url: https://staging-api.ubertruck.in/v1
    description: Staging Server
  - url: http://localhost:3000/v1
    description: Development Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
        - requestId
      properties:
        code:
          type: string
          pattern: '^[A-Z]{3}_[A-Z_]+$'
          example: 'USR_NOT_FOUND'
        message:
          type: string
          example: 'User not found with the provided ID'
        details:
          type: object
          properties:
            field:
              type: string
            reason:
              type: string
            suggestion:
              type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
        downstream:
          type: object
          properties:
            service:
              type: string
            error:
              type: string
            statusCode:
              type: integer
...
    PriceCalculation:
      type: object
      properties:
        basePrice:
          type: number
        fuelSurcharge:
          type: number
        gst:
          type: object
          properties:
            cgst:
              type: number
            sgst:
              type: number
            igst:
              type: number
            taxableAmount:
              type: number
        totalAmount:
          type: number
        priceBreakdown:
          type: object
          properties:
            ratePerKm:
              type: number
            ratePerTonne:
              type: number
            minimumCharge:
              type: number
        validUntil:
          type: string
          format: date-time
    Invoice:
      type: object
      properties:
        invoiceNumber:
          type: string
          pattern: '^INV-[0-9]{4}-[0-9]{6}$'
        bookingId:
          type: string
          format: uuid
        invoiceDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        amount:
          type: number
        gstDetails:
          type: object
          properties:
            shipperGstin:
              type: string
            carrierGstin:
              type: string
            hsnCode:
              type: string
            cgst:
              type: number
            sgst:
              type: number
            igst:
              type: number
        paymentStatus:
          type: string
          enum: [PENDING, PAID, PARTIALLY_PAID, OVERDUE]
        pdfUrl:
          type: string
          format: uri

    # Tracking Schemas
...
paths:
...
  /payments/calculate:
    post:
      tags:
        - Payments
      summary: Calculate booking price
      operationId: calculatePrice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - distance
                - weight
                - vehicleType
              properties:
                distance:
                  type: number
                  minimum: 1
                  description: Distance in kilometers
                weight:
                  type: number
                  minimum: 0.1
                  description: Weight in tonnes
                vehicleType:
                  type: string
                  enum: [TRUCK, MINI_TRUCK, TRAILER]
                cargoType:
                  type: string
                pickupPincode:
                  type: string
                deliveryPincode:
                  type: string
                surcharges:
                  type: object
                  properties:
                    fuelSurcharge:
                      type: boolean
                    nightDelivery:
                      type: boolean
                    urgentDelivery:
                      type: boolean
      responses:
        '200':
          description: Price calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceCalculation'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/invoices/{bookingId}:
    get:
      tags:
        - Payments
      summary: Get booking invoice
      operationId: getInvoice
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Payments
      summary: Generate invoice for booking
      operationId: generateInvoice
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ewayBillNumber:
                  type: string
                  pattern: '^\d{12}$'
      responses:
        '201':
          description: Invoice generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '409':
          description: Invoice already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Booking not eligible for invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...